# public/openapi.yaml
openapi: 3.0.3
info:
  title: Anikoto Scraper API
  description: |
    A high-performance REST API for scraping anime data from **anikoto.net** — built with Next.js, Cheerio, and in-memory caching.

    ## Response Format
    All endpoints return a consistent JSON envelope:
    ```json
    { "ok": true, "data": { ... }, "cached": true }
    ```
    On error:
    ```json
    { "ok": false, "message": "Error description" }
    ```

    ## Caching
    | Endpoint | TTL |
    |----------|-----|
    | `/api/home` | 5 minutes |
    | `/api/anime/:slug` | 30 minutes |
    | `/api/anime/tooltip/:id` | 30 minutes |
    | `/api/search` | 2 minutes |
    | `/api/filter` | 5 minutes |
    | `/api/schedule` | 1 hour |
    | `/api/watch/:slug` (Episodes) | 10 minutes |

    Add `?refresh=1` to any cached endpoint to bypass the cache.
  version: "1.0.0"
  contact:
    name: Teramoto
servers:
  - url: https://anikoto-scrap-api.vercel.app/api
    description: Production server
  - url: /api
    description: Local development server
...
paths:
  /home:
    get:
      tags: [Home]
      summary: Homepage data
      description: >
        Returns spotlight carousel, latest episodes, new release, and top anime by day / week / month.
        Optionally returns dynamic widget data if `widget` parameter is provided.
      operationId: getHome
      parameters:
        - name: widget
          in: query
          required: false
          schema:
            type: string
            enum: [updated-all, updated-sub, updated-dub, trending, random]
          description: Optional widget name to fetch widget data instead of full home data
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for widget data
        - $ref: '#/components/parameters/refresh'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  cached:
                    type: boolean
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/HomeData'
                      - $ref: '#/components/schemas/HomeWidgetResult'
              example:
                ok: true
                cached: false
                data:
                  spotlight: []
                  latestEpisodes: []
                  newRelease: []
                  newAdded: []
                  justCompleted: []
                  topDay: []
                  topWeek: []
                  topMonth: []
        '500':
          $ref: '#/components/responses/Error'
...
  /proxy:
    get:
      tags: [Watch]
      summary: Streaming proxy
      description: >
        High-volume zero-copy streaming proxy to bypass Cloudflare and CORS restrictions for m3u8 video streams, WebVTT subtitle files, and media segments.
        Features multi-node proxy pool routing, anti-bot browser header rotation, SSRF guard, and automatic failover retries.
        Rewrites segment URLs within m3u8 playlists to keep them routed through the proxy.
      operationId: proxyStream
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: The target m3u8 or subtitle URL to proxy
          example: "https://cdn.mewstream.buzz/.../master.m3u8"
        - name: referer
          in: query
          required: false
          schema:
            type: string
            format: uri
          description: Referer header to bypass hotlink protection
          example: "https://megaplay.buzz/"
        - name: proxy
          in: query
          required: false
          schema:
            type: string
          description: Optional custom proxy node URL (http/https/socks) for outbound routing
          example: "http://user:pass@proxy.example.com:8080"
      responses:
        '200':
          description: Proxied stream data
          content:
            application/vnd.apple.mpegurl:
              schema:
                type: string
                description: m3u8 playlist with rewritten segment URLs
            text/vtt:
              schema:
                type: string
                description: WebVTT subtitle track
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: Binary video/audio segment stream
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Access denied (SSRF guard blocked target local/private IP)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Forbidden: Access to private or local network addresses is restricted"
        '500':
          $ref: '#/components/responses/Error'
        '502':
          description: Bad Gateway (Upstream proxy failed after retries)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Proxy failed after 3 attempts"
                  detail:
                    type: string
                  url:
                    type: string
components:
  parameters:
...
    refresh:
      name: refresh
      in: query
      required: false
      schema:
        type: string
        enum: ["0", "1"]
      description: Set to `1` to bypass cache and force a fresh scrape
      example: "1"
  schemas:
...
    EpisodeStatus:
      type: object
      properties:
        sub:
          type: integer
          nullable: true
          example: 1080
        dub:
          type: integer
          nullable: true
          example: 1040
        total:
          type: integer
          nullable: true
          example: 1100
...
    LatestEpisodeItem:
      type: object
      properties:
        id:
          type: string
          example: "4127"
        slug:
          type: string
          example: "one-piece-odmau"
        title:
          type: string
          example: "One Piece"
        titleJp:
          type: string
          example: "One Piece"
        image:
          type: string
          format: uri
        href:
          type: string
          example: "/api/anime/one-piece-odmau"
        watchHref:
          type: string
          example: "https://anikoto.net/watch/one-piece-odmau/ep-1100"
        type:
          type: string
          example: "TV"
        episodes:
          $ref: '#/components/schemas/EpisodeStatus'
    TopTableItem:
      type: object
      properties:
        id:
          type: string
          example: "4127"
        slug:
          type: string
          example: "one-piece-odmau"
        title:
          type: string
          example: "One Piece"
        titleJp:
          type: string
        image:
          type: string
          format: uri
        href:
          type: string
          example: "/api/anime/one-piece-odmau"
        type:
          type: string
          example: "TV"
        episodes:
          $ref: '#/components/schemas/EpisodeStatus'
        date:
          type: string
          example: "Jul 2024"
    TopAnimeItem:
      type: object
      properties:
        rank:
          type: integer
          example: 1
        id:
          type: string
          example: "4127"
        slug:
          type: string
          example: "one-piece-odmau"
        title:
          type: string
          example: "One Piece"
        titleJp:
          type: string
        image:
          type: string
          format: uri
        href:
          type: string
          example: "/api/anime/one-piece-odmau"
        type:
          type: string
          example: "TV"
        episodes:
          $ref: '#/components/schemas/EpisodeStatus'

    # ─── Home ──────────────────────────────────────────────────────
    SpotlightAnime:
      type: object
      properties:
        slug:
          type: string
          example: "solo-leveling-season-2-arise-from-the-shadow-vvy1z"
        title:
          type: string
          example: "Solo Leveling Season 2 -Arise from the Shadow-"
        titleJp:
          type: string
        rating:
          type: string
          example: "PG-13"
        quality:
          type: string
          example: "HD"
        hasDub:
          type: boolean
          example: true
        hasSub:
          type: boolean
          example: true
        date:
          type: string
          example: "Jan 2025"
        synopsis:
          type: string
          example: "Sung Jinwoo continues his ascent as the world's strongest hunter..."
        watchUrl:
          type: string
          example: "https://anikoto.net/watch/solo-leveling-season-2-arise-from-the-shadow-vvy1z"
        href:
          type: string
          example: "/api/anime/solo-leveling-season-2-arise-from-the-shadow-vvy1z"
        image:
          type: string
          format: uri
    HomeData:
      type: object
      properties:
        spotlight:
          type: array
          items:
            $ref: '#/components/schemas/SpotlightAnime'
        latestEpisodes:
          type: array
          items:
            $ref: '#/components/schemas/LatestEpisodeItem'
        newRelease:
          type: array
          items:
            $ref: '#/components/schemas/TopTableItem'
        newAdded:
          type: array
          items:
            $ref: '#/components/schemas/TopTableItem'
        justCompleted:
          type: array
          items:
            $ref: '#/components/schemas/TopTableItem'
        topDay:
          type: array
          items:
            $ref: '#/components/schemas/TopAnimeItem'
        topWeek:
          type: array
          items:
            $ref: '#/components/schemas/TopAnimeItem'
        topMonth:
          type: array
          items:
            $ref: '#/components/schemas/TopAnimeItem'

    # ─── Anime Detail & Tooltip ────────────────────────────────────
...
    HomeWidgetResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/LatestEpisodeItem'
        currentPage:
          type: integer
          example: 1
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false
        maxPage:
          type: integer
          example: 2

    # ─── Schedule ──────────────────────────────────────────────────
...
    ApiError:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
          example: false
        message:
          type: string
          example: "Internal server error"
  responses:
    Error:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            ok: false
            message: Failed to fetch data from source
    BadRequest:
      description: Bad request — missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            ok: false
            message: Missing required parameter
...