# src/app/api/proxy/route.ts
import { NextResponse } from 'next/server';
import axios, { AxiosRequestConfig, AxiosResponse } from 'axios';
import { Readable } from 'stream';
import { proxyPool, getRandomBrowserHeaders, isPrivateUrl } from '@/lib/proxy-pool';

export const dynamic = 'force-dynamic';

// Hop-by-hop headers to strip when proxying responses
const HOP_BY_HOP_HEADERS = new Set([
  'connection',
  'keep-alive',
  'proxy-authenticate',
  'proxy-authorization',
  'te',
  'trailer',
  'transfer-encoding',
  'upgrade',
  'cf-ray',
  'cf-connecting-ip',
  'x-forwarded-for',
  'x-real-ip',
]);

export async function GET(req: Request) {
  const { searchParams, origin } = new URL(req.url);
  const targetUrl = searchParams.get('url');
  const referer = searchParams.get('referer');
  const customProxy = searchParams.get('proxy') || req.headers.get('x-proxy-target');

  if (!targetUrl) {
    return NextResponse.json({ ok: false, message: 'Missing url parameter' }, { status: 400 });
  }

  // ── SSRF Guard ─────────────────────────────────────────────────────────────
  if (isPrivateUrl(targetUrl)) {
    return NextResponse.json(
      { ok: false, message: 'Forbidden: Access to private or local network addresses is restricted' },
      { status: 403 }
    );
  }

  // ── Build Base Headers ─────────────────────────────────────────────────────
  const browserHeaders = getRandomBrowserHeaders();
  const reqHeaders: Record<string, string> = {
    'User-Agent': browserHeaders['User-Agent'],
    'Accept': '*/*',
    'Accept-Encoding': 'gzip, deflate, br',
    'Accept-Language': browserHeaders['Accept-Language'],
    'Sec-Fetch-Dest': 'empty',
    'Sec-Fetch-Mode': 'cors',
    'Sec-Fetch-Site': 'cross-site',
  };

  if (browserHeaders['sec-ch-ua']) reqHeaders['sec-ch-ua'] = browserHeaders['sec-ch-ua'];
  if (browserHeaders['sec-ch-ua-mobile']) reqHeaders['sec-ch-ua-mobile'] = browserHeaders['sec-ch-ua-mobile'];
  if (browserHeaders['sec-ch-ua-platform']) reqHeaders['sec-ch-ua-platform'] = browserHeaders['sec-ch-ua-platform'];

  if (referer) {
    reqHeaders['Referer'] = referer;
    try {
      reqHeaders['Origin'] = new URL(referer).origin;
    } catch (_) { }
  }

  // Forward Range header for partial content / video seeking support
  const rangeHeader = req.headers.get('range');
  if (rangeHeader) {
    reqHeaders['Range'] = rangeHeader;
  }

  // ── Upstream Fetch with Multi-Node Retry Logic ─────────────────────────────
  const maxAttempts = 3;
  let lastError: Error | null = null;
  let response: AxiosResponse<Readable> | null = null;
  let usedProxyUrl: string | null = null;

  for (let attempt = 1; attempt <= maxAttempts; attempt++) {
    // Select proxy node for this attempt
    usedProxyUrl = customProxy || proxyPool.getNextProxy();
    const httpsAgent = usedProxyUrl ? proxyPool.getAgent(usedProxyUrl) : undefined;

    const axiosConfig: AxiosRequestConfig = {
      headers: reqHeaders,
      timeout: 15_000,
      responseType: 'stream',
      validateStatus: (status) => status < 600,
      ...(httpsAgent ? { httpsAgent, httpAgent: httpsAgent } : {}),
    };

    try {
      const res = await axios.get<Readable>(targetUrl, axiosConfig);
      response = res;

      // If upstream returned 403 / 429 / 5xx, report failure to pool & retry next node
      if (res.status === 403 || res.status === 429 || res.status >= 500) {
        if (usedProxyUrl) proxyPool.reportFailure(usedProxyUrl);

        if (attempt < maxAttempts) {
          // Exponential backoff
          await new Promise((r) => setTimeout(r, attempt * 300));
          continue;
        }
      } else {
        // Successful response
        if (usedProxyUrl) proxyPool.reportSuccess(usedProxyUrl);
        break;
      }
    } catch (err: unknown) {
      if (usedProxyUrl) proxyPool.reportFailure(usedProxyUrl);
      lastError = err instanceof Error ? err : new Error(String(err));

      if (attempt < maxAttempts) {
        await new Promise((r) => setTimeout(r, attempt * 300));
        continue;
      }
    }
  }

  // If all retry attempts failed with exception
  if (!response) {
    console.error(`[Proxy Error] All ${maxAttempts} attempts failed for ${targetUrl}:`, lastError);
    return NextResponse.json(
      { ok: false, message: `Proxy failed after ${maxAttempts} attempts`, detail: lastError?.message, url: targetUrl },
      { status: 502 }
    );
  }

  // Forward upstream block or failure status if final attempt still blocked
  if (response.status === 403 || response.status === 401) {
    console.error(`[Proxy] Upstream blocked (${response.status}) on ${targetUrl}`);
    return NextResponse.json(
      {
        ok: false,
        message: `Upstream server blocked the request (HTTP ${response.status}). Try using the Cloudflare Worker proxy instead.`,
        upstreamStatus: response.status,
        url: targetUrl,
      },
      { status: response.status }
    );
  }

  if (response.status >= 400) {
    return NextResponse.json(
      { ok: false, message: `Upstream error: HTTP ${response.status}`, upstreamStatus: response.status, url: targetUrl },
      { status: response.status }
    );
  }

  // ── Prepare Response Headers ───────────────────────────────────────────────
  const resHeaders = new Headers();
  const contentType = (response.headers['content-type'] as string) || '';

  // Copy non-hop-by-hop response headers
  Object.entries(response.headers).forEach(([key, val]) => {
    const lowerKey = key.toLowerCase();
    if (!HOP_BY_HOP_HEADERS.has(lowerKey) && val !== undefined) {
      resHeaders.set(key, Array.isArray(val) ? val.join(', ') : String(val));
    }
  });

  resHeaders.set('Access-Control-Allow-Origin', '*');
  resHeaders.set('Access-Control-Allow-Methods', 'GET, OPTIONS');
  resHeaders.set('Access-Control-Allow-Headers', '*');
  resHeaders.set('X-Accel-Buffering', 'no'); // Disable reverse proxy buffering

  const isManifest = /\.m3u8/i.test(targetUrl) || contentType.includes('mpegurl') || contentType.includes('m3u8');
  const isSubtitle = /\.(vtt|srt|ass)$/i.test(targetUrl) || contentType.includes('vtt');

  // ── Subtitle Response ──────────────────────────────────────────────────────
  if (isSubtitle) {
    resHeaders.set('Content-Type', 'text/vtt; charset=utf-8');
    resHeaders.set('Cache-Control', 'public, max-age=3600');
    const webStream = Readable.toWeb(response.data) as ReadableStream<Uint8Array>;
    return new Response(webStream, { status: response.status, headers: resHeaders });
  }

  // ── HLS Playlist Rewriting ────────────────────────────────────────────────
  if (isManifest) {
    resHeaders.set('Content-Type', 'application/vnd.apple.mpegurl');
    resHeaders.set('Cache-Control', 'no-store, no-cache, must-revalidate');

    // Read full playlist text
    const chunks: Uint8Array[] = [];
    const stream = response.data;
    for await (const chunk of stream) {
      chunks.push(chunk);
    }
    const text = Buffer.concat(chunks).toString('utf-8');
    const baseUrl = new URL(targetUrl);

    // Build base proxy URL path
    const proxyPath = `${origin}/api/proxy`;

    const rewrittenText = text
      .split('\n')
      .map((line) => {
        // Rewrite URI attributes in HLS tags (keys, maps, subtitles)
        if (line.includes('URI=')) {
          line = line.replace(/URI=["']([^"']+)["']/g, (match, uri) => {
            try {
              let keyUrl = uri.startsWith('http') ? uri : new URL(uri, baseUrl).toString();

              // Domain fixup for CDN mirrors
              try {
                const parsedKey = new URL(keyUrl);
                if (
                  parsedKey.hostname.endsWith('.buzz') ||
                  parsedKey.hostname.endsWith('.click') ||
                  parsedKey.hostname.includes('zaplume.buzz') ||
                  parsedKey.hostname.includes('mewstream.buzz')
                ) {
                  parsedKey.host = baseUrl.host;
                  keyUrl = parsedKey.toString();
                }
              } catch (_) { }

              let proxied = `${proxyPath}?url=${encodeURIComponent(keyUrl)}`;
              if (referer) proxied += `&referer=${encodeURIComponent(referer)}`;
              if (customProxy) proxied += `&proxy=${encodeURIComponent(customProxy)}`;
              return `URI="${proxied}"`;
            } catch {
              return match;
            }
          });
        }

        const trimmed = line.trim();
        if (!trimmed || trimmed.startsWith('#')) return line;

        // Media segment or sub-playlist line
        try {
          let segmentUrl = trimmed.startsWith('http') ? trimmed : new URL(trimmed, baseUrl).toString();

          // Domain fixup for CDN mirrors
          try {
            const parsedSeg = new URL(segmentUrl);
            if (
              parsedSeg.hostname.endsWith('.buzz') ||
              parsedSeg.hostname.endsWith('.click') ||
              parsedSeg.hostname.includes('zaplume.buzz') ||
              parsedSeg.hostname.includes('mewstream.buzz')
            ) {
              parsedSeg.host = baseUrl.host;
              segmentUrl = parsedSeg.toString();
            }
          } catch (_) { }

          let proxied = `${proxyPath}?url=${encodeURIComponent(segmentUrl)}`;
          if (referer) proxied += `&referer=${encodeURIComponent(referer)}`;
          if (customProxy) proxied += `&proxy=${encodeURIComponent(customProxy)}`;
          return proxied;
        } catch {
          return line;
        }
      })
      .join('\n');

    return new Response(rewrittenText, {
      status: response.status,
      headers: resHeaders,
    });
  }

  // ── Video / Audio Binary Streaming (Zero-Copy ReadableStream) ────────────
  const isRealMedia =
    contentType.includes('video') ||
    contentType.includes('audio') ||
    contentType.includes('octet-stream') ||
    contentType.includes('mp4') ||
    contentType.includes('mpeg');

  resHeaders.set('Content-Type', isRealMedia ? contentType : 'application/octet-stream');
  resHeaders.set('Cache-Control', 'public, max-age=7200, immutable');

  // Convert Node readable stream to Web ReadableStream for zero-copy streaming
  const webStream = Readable.toWeb(response.data) as ReadableStream<Uint8Array>;

  return new Response(webStream, {
    status: response.status === 206 ? 206 : response.status,
    headers: resHeaders,
  });
}